apps/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.4,<3.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
httpx>=0.28,<1.0
pytest>=8.3,<9.0
